project_18_simple_rag_bot/requirements.txt
langchain
langchain-google-genai
python-dotenv
faiss-cpu
langchain-community
uv